Separated all inputs into separate scripts and cleaned up the code.
Right now the mouse look is not working as smooth as I would hoped, and will look into this at some point in the future. As of right now, it appears to be an issue with the Input System, so till I know more, not sure thee is much I can do just yet.